// projects/iblopay-admin/src/app/modules/transactions/enums/transaction-type.enum.d.ts
export declare enum TransactionType {
    DEPOSIT = "DEPOSIT",
    WITHDRAWAL = "WITHDRAWAL",
    TRANSFER = "TRANSFER",
    PAYMENT_NFC = "PAYMENT_NFC",
    SWEEP = "SWEEP",
    SWEEP_INVERSE = "SWEEP_INVERSE",
    COMMISSION = "COMMISSION",
    REIMBURSEMENT = "REIMBURSEMENT"
}
//# sourceMappingURL=transaction-type.enum.d.ts.map